% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-dll.R
\name{clean_dll}
\alias{clean_dll}
\title{Remove compiled objects from /src/ directory}
\usage{
clean_dll(path = ".")
}
\arguments{
\item{path}{Path to a package, or within a package.}
}
\description{
Invisibly returns the names of the deleted files.
}
\seealso{
\code{\link[=compile_dll]{compile_dll()}}
}
