% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-tools.R
\name{has_build_tools}
\alias{check_build_tools}
\alias{has_build_tools}
\title{Are build tools are available?}
\usage{
has_build_tools(debug = FALSE)

check_build_tools(debug = FALSE)
}
\arguments{
\item{debug}{If \code{TRUE}, will print out extra information useful for
debugging.}
}
\description{
\code{has_build_tools} returns a logical, \code{check_build_tools} throws
an error. If run interactively from RStudio, and the built-tools are not
available either function will trigger an automated install.
}
\examples{
has_build_tools()
check_build_tools()
}

