% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile-dll.R
\name{needs_compile}
\alias{needs_compile}
\title{Does the package need recompiling?
(i.e. is there a source or header file newer than the dll)}
\usage{
needs_compile(path = ".")
}
\arguments{
\item{path}{Path to a package, or within a package.}
}
\description{
Does the package need recompiling?
(i.e. is there a source or header file newer than the dll)
}
\keyword{internal}
