% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_src.R
\name{pkg_has_src}
\alias{pkg_has_src}
\title{Does a source package have \code{src/} directory?}
\usage{
pkg_has_src(path = ".")
}
\arguments{
\item{path}{Path to package (or directory within package).}
}
\description{
If it does, you definitely need build tools.
}
