% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c-registration.R
\name{pkg_links_to_rcpp}
\alias{pkg_links_to_rcpp}
\alias{pkg_links_to_cpp11}
\title{Test if a package path is linking to Rcpp or cpp11}
\usage{
pkg_links_to_rcpp(path)

pkg_links_to_cpp11(path)
}
\arguments{
\item{path}{Path to a package, or within a package.}
}
\description{
Test if a package path is linking to Rcpp or cpp11
}
\keyword{internal}
