% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgbuild-package.R
\docType{package}
\name{pkgbuild-package}
\alias{pkgbuild}
\alias{pkgbuild-package}
\title{pkgbuild: Find Tools Needed to Build R Packages}
\description{
Provides functions used to build R packages. Locates compilers needed to build R packages on various platforms and ensures the PATH is configured appropriately so R can use them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/pkgbuild}
  \item \url{https://pkgbuild.r-lib.org}
  \item Report bugs at \url{https://github.com/r-lib/pkgbuild/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Hadley Wickham
  \item Jim Hester
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
