% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-bg.R
\name{pkgbuild_process}
\alias{pkgbuild_process}
\title{Build package in the background}
\description{
This R6 class is a counterpart of the \code{\link[=build]{build()}} function, and
represents a background process that builds an R package.
}
\section{Usage}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{bp <- pkgbuild_process$new(path = ".", dest_path = NULL,
         binary = FALSE, vignettes = TRUE, manual = FALSE, args = NULL)
bp$get_dest_path()
}\if{html}{\out{</div>}}

Other methods are inherited from \link[callr:rcmd_process]{callr::rcmd_process} and
\code{processx::process}.
}

\section{Arguments}{

See the corresponding arguments of \code{\link[=build]{build()}}.
}

\section{Details}{

Most methods are inherited from \link[callr:rcmd_process]{callr::rcmd_process} and
\code{processx::process}.

\code{bp$get_dest_path()} returns the path to the built package.
}

\section{Examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{## Here we are just waiting, but in a more realistic example, you
## would probably run some other code instead...
bp <- pkgbuild_process$new("mypackage", dest_path = tempdir())
bp$is_alive()
bp$get_pid()
bp$wait()
bp$read_all_output_lines()
bp$read_all_error_lines()
bp$get_exit_status()
bp$get_dest_path()
}\if{html}{\out{</div>}}
}

