% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmd.R
\name{rcmd_build_tools}
\alias{rcmd_build_tools}
\title{Call R CMD <command> with build tools active}
\usage{
rcmd_build_tools(..., required = TRUE)
}
\arguments{
\item{...}{Parameters passed on to \code{rcmd_safe}.}
}
\description{
This is a wrapper around \code{\link[callr]{rcmd_safe}} that checks
that you have build tools available, and on Windows, automatically sets
the path to include Rtools.
}
\examples{
if (has_build_tools()) {
  rcmd_build_tools("CONFIG", "CC")$stdout
  rcmd_build_tools("CC", "--version")$stdout
}
}

