% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtools.R
\name{rtools_needed}
\alias{rtools_needed}
\title{Retrieve a text string with the rtools version needed}
\usage{
rtools_needed(r_version = getRversion())
}
\description{
Retrieve a text string with the rtools version needed
}
\keyword{internal}
