% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-debug.R
\name{with_no_compiler}
\alias{with_no_compiler}
\title{Temporarily break code compilation}
\usage{
with_no_compiler(code)
}
\arguments{
\item{code}{Code to execute with broken compilers}
}
\description{
This is useful for testing.
}

