% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with-debug.R
\name{without_compiler}
\alias{without_compiler}
\alias{without_cache}
\title{Tools for testing pkgbuild}
\usage{
without_compiler(code)

without_cache(code)
}
\arguments{
\item{code}{Code to execute with broken compilers}
}
\description{
\code{with_compiler} temporarily disables code compilation by setting
\code{CC}, \code{CXX}, makevars to \code{test}. \code{without_cache}
resets the cache before and after running \code{code}.
}
