% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-tools.R
\name{has_build_tools}
\alias{check_build_tools}
\alias{has_build_tools}
\alias{with_build_tools}
\title{Are build tools are available?}
\usage{
has_build_tools(debug = FALSE)

check_build_tools(debug = FALSE)

with_build_tools(code, debug = FALSE, required = TRUE)
}
\arguments{
\item{debug}{If \code{TRUE}, will print out extra information useful for
debugging. If \code{FALSE}, it will use result cached from a previous run.}

\item{code}{Code to rerun in environment where build tools are guaranteed to
exist.}

\item{required}{If \code{TRUE}, and build tools are not available,
will throw an error. Otherwise will attempt to run \code{code} without
them.}
}
\description{
\code{has_build_tools} returns a logical, \code{check_build_tools} throws
an error. \code{with_build_tools} checks that build tools are available,
then runs \code{code} in an correctly staged environment.
If run interactively from RStudio, and the build tools are not
available these functions will trigger an automated install.
}
\examples{
has_build_tools()
check_build_tools()
}

