% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp-attributes.R
\name{pkg_links_to_rcpp}
\alias{pkg_links_to_rcpp}
\title{Test if a package path is linking to Rcpp}
\usage{
pkg_links_to_rcpp(path)
}
\arguments{
\item{path}{Path to a package, or within a package.}
}
\description{
Test if a package path is linking to Rcpp
}
\keyword{internal}
